Tweak ViewEntry dialog for formatting and wordwrap.
Tweak the grid layout of the ViewEntry dialog so that the notes entry
expands horizontally and vertically upon window resize.  Additionally
change determination of -wraplength for the notes entry to be based
upon window width and adjust so that the -wraplength is dynamically
determined based upon current window width value.