Change code to address github issue #25 "Can't Add Login"
Changes adjust how Gorilla handles "add login" function when an existing
login is selected in the treeview.  The new code, when an add login action
is performed while an existing login is selected in the tree, will perform
the add login action using the group membership of the selected login.