Changes resulting from Issue 53 report regarding arbitrary limit on key stretching iterations count.
Removed the arbitrary limit check.

Adjusted password db preferences to allow up to 2^31-1 as the maximum
iteration count.  This is the maximum value that Gorilla can read/write into
the iterations field of the password safe v3 format.  It is also a large
enough value that it should be quite some time before CPU/GPU performance
overtakes this upper limit.

Edited help.text file to better explain the meaning of the value, and to
give a small hint as to how to pick a somewhat reasonable iteration value.
Semi-automatic selection of a time delay based iteration value by
Gorilla itself will be the subject of a follow-on commit.

Added an actual error message for failure to scan an iteration value from
the pwsafe binary data instead of just "oops".