Fix csv-export and backup-timestamp unit tests.
Fix csv-export to set necessary export options before performing export so
that actual and expected values match properly.

Fix race condition in timestamp backup test.  Depending on the exact timing
of the run, the actual and expected timestamps sometimes differed by one
second, resulting in a false failure.  Change to simply look for presence of
a file with the correct filename format in a brand new temporary directory.