Move storage, access, and updating of file HMAC into Ictl object.
This fixes a bug that only exhibited during a merge operation where the
second DB that was merged improperly overwrote the HMAC of the main
loaded DB.  This then produced the effect of a fake change to the main
loaded DB.  Storing the HMAC in the Itcl object provides independent
storage for each opened DB, preventing any fake change notices due to
the original overwrite.

Note, this commit only works with V3 PasswordSafe files.