Code to fix issue #194 - nautilus file manager breaks gorilla auto-copy
Adds a new preference value that holds a latency value.  Any X11 paste
events that occur within a time window of this value's length beginning
at the start of an auto-copy operation continue to insert the username
into the clipboard.

Testing on the author's system shows promise, as this change prevents
pasting of the username followed immediately by the password when
running the xclipboard utility.  A latency as low as 20ms seemed to
work on the author's system, so a default of 50ms was chosen to leave a
little slack for different system's performance factors.  The default
of 50ms should also be slower than even the fastest human reaction time
for commanding a paste operation using either the mouse or the
keyboard.  However the value is user adjustable to allow for personal
tweaking as necessary.